Revert "fix(mobile-native-chat): retire pending bubbles glued into one transc…"
This reverts commit 68ca17e46c6dad0ffda2f8d593ad1a434520f346.
