+++ mobile/src/session/mobile-native-chat-pending-echo.ts
@@ -2,8 +2,6 @@ export type MobileNativeChatPendingMessage = {
   id: string
   text: string
   expectedOccurrence: number
-  /** Whether the transcript snapshot captured before this send was authoritative. */
-  glueBaselineTrusted: boolean
   /** Local preview URIs carried by the send for its optimistic echo. */
   images?: string[]
   baselineTailMessageId: string | null
@@ -15,7 +13,6 @@ export type MobileNativeChatSendOrigin = {
   normalizedText: string
   baselineOccurrences: number
   baselineTailMessageId: string | null
-  glueBaselineTrusted: boolean
 }
 
 type PendingByKey = Record<string, MobileNativeChatPendingMessage[]>
@@ -59,7 +56,6 @@ export function appendMobileNativeChatPending(
             ? expectedImageEchoOrdinal
             : origin.baselineOccurrences + earlierOutstanding + 1,
         baselineTailMessageId: origin.baselineTailMessageId,
-        glueBaselineTrusted: origin.glueBaselineTrusted,
         ...(images?.length ? { images } : {})
       }
     ]
