+++ mobile/src/session/use-mobile-native-chat-drafts.ts
@@ -1,15 +1,16 @@
 import { useCallback, useEffect, useRef, useState, type Dispatch, type SetStateAction } from 'react'
 import type { NativeChatMessage } from '../../../src/shared/native-chat-types'
 import {
+  countImageSourceTurnsAfter,
   countUserTextOccurrences,
   findLandedImagePreviewEchoes,
   findLandedUnconfirmedSends,
   mergeLandedImagePreviewEchoes,
   migrateImagePreviewMessageIds,
   normalizeReconcileText,
+  normalizedUserText,
   type UnconfirmedSend
 } from './mobile-native-chat-draft-reconcile'
-import { retireLandedMobileNativeChatPending } from './mobile-native-chat-pending-retirement'
 import {
   appendMobileNativeChatPending,
   combineMobileNativeChatPending,
@@ -137,11 +138,10 @@ export function useMobileNativeChatDrafts(args: {
         pendingKey,
         normalizedText,
         baselineOccurrences: countUserTextOccurrences(messagesRef.current, normalizedText),
-        baselineTailMessageId: messagesRef.current.at(-1)?.id ?? null,
-        glueBaselineTrusted: !transcriptLoading
+        baselineTailMessageId: messagesRef.current.at(-1)?.id ?? null
       }
     },
-    [draftKey, pendingKey, transcriptLoading]
+    [draftKey, pendingKey]
   )
 
   // Why: over relay the send RPC can take seconds (or lose only its ack), and a
@@ -291,7 +291,27 @@ export function useMobileNativeChatDrafts(args: {
     }
     setPendingBySession((previous) => {
       const current = previous[pendingKey] ?? []
-      const next = retireLandedMobileNativeChatPending(messages, current, landedImagePendingIds)
+      const landedCounts = new Map<string, number>()
+      for (const message of messages) {
+        const text = normalizedUserText(message)
+        if (text) {
+          landedCounts.set(text, (landedCounts.get(text) ?? 0) + 1)
+        }
+      }
+      // Image-only source-turn counts stay stable across reruns and ignore paginated history.
+      const next = current.filter((item) => {
+        if (landedImagePendingIds.has(item.id)) {
+          return false
+        }
+        // Keep image echoes until their local preview reaches the authoritative message.
+        if (item.images?.length) {
+          return true
+        }
+        return item.text.trim() === ''
+          ? countImageSourceTurnsAfter(messages, item.baselineTailMessageId) <
+              item.expectedOccurrence
+          : (landedCounts.get(normalizeReconcileText(item.text)) ?? 0) < item.expectedOccurrence
+      })
       if (next.length === current.length) {
         return previous
       }
