+++ mobile/src/session/use-mobile-native-chat-message-send.test.ts
@@ -36,7 +36,6 @@ describe('useMobileNativeChatMessageSend', () => {
   let renderer: ReactTestRenderer | null = null
   let api: Send | null = null
   const acceptSend = vi.fn()
-  const captureSendOrigin = vi.fn(() => ({ draftKey: 'k', pendingKey: 'p' }) as never)
   const holdUnconfirmedSend = vi.fn()
   const onCommandSend = vi.fn()
   const commandSendRef = { current: onCommandSend }
@@ -56,7 +55,7 @@ describe('useMobileNativeChatMessageSend', () => {
         deviceTokenRef: { current: 'device' },
         agentRef,
         commandSendRef,
-        captureSendOrigin,
+        captureSendOrigin: () => ({ draftKey: 'k', pendingKey: 'p' }) as never,
         readSeededLaunchDraftSeed,
         clearDraftForSend: () => {},
         restoreRejectedDraft: () => {},
@@ -72,12 +71,10 @@ describe('useMobileNativeChatMessageSend', () => {
   }
 
   const sentArgs = (): {
-    text?: string
     clearInputFirst?: boolean
     resolvedLaunchDraft?: { text: string; createdAt: number }
   } =>
     sendWithOutcome.mock.calls[0]![0] as {
-      text?: string
       clearInputFirst?: boolean
       resolvedLaunchDraft?: { text: string; createdAt: number }
     }
@@ -92,7 +89,6 @@ describe('useMobileNativeChatMessageSend', () => {
     typeCommandWithOutcome.mockReset()
     typeCommandWithOutcome.mockResolvedValue('accepted')
     acceptSend.mockReset()
-    captureSendOrigin.mockClear()
     holdUnconfirmedSend.mockReset()
     onCommandSend.mockReset()
     commandSendRef.current = onCommandSend
@@ -384,45 +380,5 @@ describe('useMobileNativeChatMessageSend', () => {
     // The answer released its own hold on the way out.
     expect(acquireMobileNativeChatTerminalWrite('term')).toBe(true)
     releaseMobileNativeChatTerminalWrite('term')
-  })
-
-  // The host writes these bytes verbatim, so whitespace the match key drops must
-  // not reach the agent's input line and glue the next rapid send onto it (#14262).
-  it('trims trailing whitespace without changing leading prompt content', async () => {
-    mount(() => null)
-    await act(async () => {
-      await api!.send('  run the tests \n')
-    })
-    expect(sentArgs().text).toBe('  run the tests')
-    expect(captureSendOrigin).toHaveBeenCalledWith('  run the tests')
-    expect(acceptSend.mock.calls[0]![1]).toBe('  run the tests')
-  })
-
-  it('trims trailing whitespace on a question answer too', async () => {
-    mount(() => null)
-    await act(async () => {
-      await api!.answerQuestion('\n 1 ')
-    })
-    expect(sentArgs().text).toBe('\n 1')
-  })
-
-  it('keeps a leading-whitespace slash draft as prose', async () => {
-    mount(() => null, 'codex')
-    await act(async () => {
-      await api!.send(' /delete ')
-    })
-    expect(sentArgs().text).toBe(' /delete')
-    expect(sendWithOutcome).toHaveBeenCalledOnce()
-    expect(typeCommandWithOutcome).not.toHaveBeenCalled()
-    expect(acceptSend).toHaveBeenCalledWith(expect.anything(), ' /delete', undefined)
-    expect(onCommandSend).not.toHaveBeenCalled()
-  })
-
-  it('keeps interior newlines of a multi-line prompt', async () => {
-    mount(() => null)
-    await act(async () => {
-      await api!.send('first line\nsecond line\n')
-    })
-    expect(sentArgs().text).toBe('first line\nsecond line')
   })
 })
